Fix bug.

+++ database/migrations/2024_10_02_081345_create_properties_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             $table->string('name');
 
-            $table->integer('attribute_id');
+            $table->unsignedBigInteger('attribute_id');
             $table->foreign('attribute_id')->references('id')->on('attributes')->onDelete('cascade');
 
             $table->string('status')->default(\App\Enums\PropertiStatus::ACTIVE);
